+++ src/executors/executor.ts
@@ -44,6 +44,7 @@ import { DataStore } from '../processors/DataStore';
 import { createClientLogger } from '../logger';
 import { getCurrentEnvironment } from '../environments';
 import { SwapConnectorImplementationService } from '../processors/swap-connector-implementation.service';
+import { tryInitTakerALT } from '../providers/utils/tryInitAltSolana';
 
 export type ExecutorInitializingChain = Readonly<{
   chain: ChainId;
@@ -100,82 +101,6 @@ export type ExecutorSupportedChain = Readonly<{
   fulfillProvider: ProviderAdapter;
   beneficiary: Address;
 }>;
-
-export async function waitSolanaTxFinalized(connection: Connection, txId: string) {
-  let finalized = false;
-  while (!finalized) {
-    // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-    const result = await connection.getTransaction(txId, {
-      commitment: 'finalized',
-      maxSupportedTransactionVersion: 1,
-    });
-    if (result !== null) {
-      finalized = true;
-      break;
-    }
-    // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-    await helpers.sleep(1000);
-  }
-}
-
-export async function tryInitTakerALT(
-  wallet: helpers.Wallet,
-  chains: ChainId[],
-  solanaClient: Solana.DlnClient,
-  logger: Logger,
-  retries = 3,
-) {
-  for (let i = 0; i < retries; i += 1) {
-    // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-    const maybeTxs = await solanaClient.initForTaker(wallet.publicKey, chains);
-    if (!maybeTxs) {
-      logger.info(
-        `ALT already initialized or was found: ${solanaClient.fulfillPreswapALT!.toBase58()}`,
-      );
-
-      return;
-    }
-
-    const solanaConnection = solanaClient.getConnection(ChainId.Solana);
-    try {
-      const [initTx, ...restTxs] = maybeTxs;
-      // initALT ix may yield errors like recentSlot is too old/broken blockhash, it's better to add retries
-      // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-      const [initTxId] = await helpers.sendAll(solanaConnection, wallet, initTx, {
-        convertIntoTxV0: false,
-        blockhashCommitment: 'confirmed',
-      });
-      logger.info(`Initialized ALT: ${initTxId}`);
-
-      // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-      await waitSolanaTxFinalized(solanaConnection, initTxId);
-      const txWithFreeze = restTxs.pop();
-      if (restTxs.length !== 0) {
-        // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-        const fillIds = await helpers.sendAll(solanaConnection, wallet, restTxs, {
-          convertIntoTxV0: false,
-          blockhashCommitment: 'confirmed',
-        });
-        // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-        await Promise.all(fillIds.map((txId) => waitSolanaTxFinalized(solanaConnection, txId)));
-        logger.info(`Fill ALT: ${fillIds.join(', ')}`);
-      }
-      if (txWithFreeze) {
-        // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-        const [freezeId] = await helpers.sendAll(solanaConnection, wallet, txWithFreeze, {
-          convertIntoTxV0: false,
-          blockhashCommitment: 'confirmed',
-        });
-        logger.info(`Freezed ALT: ${freezeId}`);
-      }
-
-      return;
-    } catch (e) {
-      logger.error(e);
-    }
-  }
-  throw new Error('Failed to init ALT');
-}
 
 export interface IExecutor {
   readonly tokenPriceService: PriceTokenService;
@@ -396,18 +321,6 @@ export class Executor implements IExecutor {
         await client.destination.debridge.init();
 
         // TODO: wait until solana enables getProgramAddress with filters for ALT and init ALT if needed
-
-        /* const altInitTx = await client.initForTaker(
-          new PublicKey(client.parseAddress(chain.beneficiary)),
-          config.chains.map((chainConfig) => chainConfig.chain),
-        );
-        if (altInitTx) {
-          this.logger.info(`Initializing Solana Address Lookup Table (ALT)`);
-          // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
-          await fulfillProvider.sendTransaction(altInitTx, { logger: this.logger });
-        } else {
-          this.logger.info(`Solana Address Lookup Table (ALT) already exists`);
-        } */
 
         // eslint-disable-next-line no-await-in-loop -- Intentional because works only during initialization
         await tryInitTakerALT(
